Organization, extended +/- trend lines
Tried to remove some hard coding. Modularized the code a bit to make it
easier to change the variables to show a week long summary as opposed
to just 24hr summary. More work needs to be done there.
Also added extra curveVertex points to start and ends of the line so
that it would show the line for the first and last buckets of the
histogram.