Added function to display percentage 'productive' time.
This is different from the abs(plus/minus) display of productive v
non-productive time. Intent is to show what percentage of time spent in
a given interval (histogram bucket) is productive.
Its a bit confusing that the line rendered as it is right now overlaps
the bar chart/points/lines but that can be fixed (additional axis?
should be something more elegant).